+++ crates/bitwarden-vault/src/folder/folder_models.rs
