+++ crates/bitwarden-vault/src/folder/mod.rs
@@ -1,0 +1,7 @@
+mod folder_client;
+mod folder_create;
+mod folder_models;
+
+pub use folder_client::*;
+pub use folder_create::*;
+pub use folder_models::*;
