+++ crates/bitwarden-vault/src/lib.rs
@@ -12,9 +12,7 @@ pub use collection::{Collection, CollectionView};
 mod collection_client;
 pub use collection_client::CollectionsClient;
 mod folder;
-pub use folder::{Folder, FolderView};
-mod folder_client;
-pub use folder_client::FoldersClient;
+pub use folder::*;
 mod password_history;
 pub use password_history::{PasswordHistory, PasswordHistoryView};
 mod password_history_client;
